requirements.txt
python-telegram-bot==20.3
requests
pandas
scikit-learn
nba_api